examples/python-fastapi/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
jinja2>=3.1
python-multipart>=0.0.9
httpx>=0.27
python-dotenv>=1.0
authon
